extracted Download Task as "Downloader" class
